Users should only be able to edit their own facilities.
Non Admin users do not see a link to edit their own facilities. The /edit path still works, making this security merely cosmetic since any user can access the edit page for any facility. 

We have a facility_users link table we can use to filter access. Perhaps the table needs to be updated to include a facility role as well. Roles could be 'admin', 'owner', 'staff'. We can use these roles to enable/disable certain behaviors per facilities in the future. 

Even without the extra role, we can still use the table to keep non facility users out of all but the "show" action. 

This change implements the access filter for facilities. Access is restricted for anything other than the "show" action for non facility user. 

If we think this is a reasonable approach, I can keep going down this path and improve this to include the roles as well.